Update base_tuner.py
Hi,
I tried to convert the output from this function results_summary(), but it returns None because it is missing a return command at the end. To resolve this issue, I suggest adding a return statement at the end of this function as follows: "return best_trials". This way, the results can be passed into a variable and be converted to a Dataframe or other formats.